cleanup

+++ HealthEntry/Items/HealthEntryItemManager.h
@@ -28,7 +28,7 @@
 /**
  * @return The number of selected items.
  */
-- (int)countOfSelectedItems;
+- (NSUInteger)countOfSelectedItems;
 
 /**
  * Selects the given item for data entry.
